Move SimpleScratchpad
