+++ platform-sdk/swirlds-platform-core/src/testFixtures/java/module-info.java
@@ -16,10 +16,8 @@ open module com.swirlds.platform.core.test.fixtures {
     requires transitive org.hiero.consensus.model;
     requires transitive org.hiero.consensus.platformstate;
     requires transitive org.hiero.consensus.state;
-    requires transitive org.hiero.consensus.utility;
     requires com.swirlds.base.test.fixtures;
     requires com.swirlds.config.extensions;
-    requires com.swirlds.logging;
     requires com.swirlds.merkledb.test.fixtures;
     requires com.swirlds.merkledb;
     requires com.swirlds.state.impl.test.fixtures;
@@ -29,6 +27,7 @@ open module com.swirlds.platform.core.test.fixtures {
     requires org.hiero.consensus.reconnect;
     requires org.hiero.consensus.roster.test.fixtures;
     requires org.hiero.consensus.utility.test.fixtures;
+    requires org.hiero.consensus.utility;
     requires com.github.spotbugs.annotations;
     requires org.junit.jupiter.api;
     requires org.mockito;
